Move time travel func, add code split feat.
