+++ packages/alpress-contracts/.npmignore
@@ -1,0 +1,1 @@
+src/
