+++ packages/dlog-utils/.npmignore
@@ -1,0 +1,4 @@
+src
+
+.nyc_output
+tsconfig.*
