+++ packages/dlog-utils/src/index.ts
@@ -1,5 +1,3 @@
-import {localSetup} from './local-setup';
-import {timeTravel} from './local-setup';
-import {loadJSON} from './load-json';
-
-export {localSetup, timeTravel, loadJSON };
+// export { localSetup } from './local-setup';
+export { timeTravel } from './time-travel';
+export { loadJSON } from './load-json';
