+++ packages/dlog/.npmignore
@@ -1,15 +1,17 @@
 src
 repo
 test
-tsconfig.json
-tsconfig.module.json
-tslint.json
-.travis.yml
-.github
-.prettierignore
-.vscode
+
 build/docs
 **/*.spec.*
 coverage
-.nyc_output
+
+tsconfig.*
+.eslintrc.js
+
 *.log
+.github
+.nyc_output
+.prettierignore
+.travis.yml
+.vscode
