+++ packages/dlog/src/lib/models/index.ts
@@ -1,17 +1,7 @@
-import { Article } from './article';
-import { ArticleHeader } from './article-header';
-import { ArticlesIndex } from './articles-index';
-import { Author } from './author';
-import { Bucket } from './bucket';
-import { ENSContent } from './ens-content';
-import { Identity } from './identity';
-
-export {
-  Article,
-  ArticleHeader,
-  ArticlesIndex,
-  Author,
-  Bucket,
-  ENSContent,
-  Identity
-};
+export { Article } from './article';
+export { ArticleHeader } from './article-header';
+export { ArticlesIndex } from './articles-index';
+export { Author } from './author';
+export { Bucket } from './bucket';
+export { ENSContent } from './ens-content';
+export { Identity } from './identity';
