+++ packages/dlog/tsconfig.module.json
@@ -6,6 +6,7 @@
     "module": "esnext"
   },
   "exclude": [
-    "node_modules/**"
+    "node_modules/**",
+    "src/**/*.spec.ts"
   ]
 }
